+++ src/main/java/guideme/compiler/IndexingContext.java
@@ -5,10 +5,16 @@ import guideme.extensions.Extension;
 import guideme.extensions.ExtensionCollection;
 import guideme.extensions.ExtensionPoint;
 import guideme.indices.PageIndex;
+import guideme.libs.mdast.model.MdAstAnyContent;
 import java.util.List;
 import net.minecraft.resources.ResourceLocation;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 
+/**
+ * The context used during search indexing of custom tags {@link TagCompiler}.
+ */
+@ApiStatus.NonExtendable
 public interface IndexingContext {
     ExtensionCollection getExtensions();
 
@@ -22,6 +28,14 @@ public interface IndexingContext {
     ResourceLocation getPageId();
 
     PageCollection getPageCollection();
+
+    default void indexContent(List<? extends MdAstAnyContent> children, IndexingSink sink) {
+        for (var child : children) {
+            indexContent(child, sink);
+        }
+    }
+
+    void indexContent(MdAstAnyContent content, IndexingSink sink);
 
     default byte @Nullable [] loadAsset(ResourceLocation imageId) {
         return getPageCollection().loadAsset(imageId);