+++ src/main/java/guideme/internal/screen/GuideNavigation.java
@@ -2,16 +2,23 @@ package guideme.internal.screen;
 
 import guideme.Guide;
 import guideme.PageAnchor;
+import java.net.URI;
 import java.util.Objects;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.ConfirmLinkScreen;
 import net.minecraft.client.gui.screens.Screen;
 import org.jetbrains.annotations.Nullable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Extracts the history navigation logic from GuideScreen to allow for jumping between search and guide display
  * seamlessly without duplicating all the nav logic.
  */
 public final class GuideNavigation {
+    private static final Logger LOG = LoggerFactory.getLogger(GuideNavigation.class);
+
     private GuideNavigation() {
     }
 
@@ -65,5 +72,34 @@ public final class GuideNavigation {
     public static void navigateBack(Guide guide) {
         var history = GlobalInMemoryHistory.get(guide);
         history.back().ifPresent(pageAnchor -> navigateTo(guide, pageAnchor));
+    }
+
+    public static void openUrl(String href) {
+        URI uri;
+        try {
+            uri = URI.create(href);
+        } catch (IllegalArgumentException ignored) {
+            LOG.debug("Can't parse '{}' as URL", href);
+            return;
+        }
+
+        // Treat it as an external URL if it has a scheme
+        var minecraft = Minecraft.getInstance();
+        var previousScreen = minecraft.screen;
+
+        if (uri.getScheme() != null) {
+            if (minecraft.options.chatLinksPrompt().get().booleanValue()) {
+                minecraft.setScreen(new ConfirmLinkScreen(doOpen -> {
+                    if (doOpen) {
+                        Util.getPlatform().openUri(uri);
+                    }
+                    minecraft.setScreen(previousScreen);
+                }, href, false));
+            } else {
+                Util.getPlatform().openUri(uri);
+            }
+        } else {
+            LOG.debug("Can't open relative URL: '{}'", href);
+        }
     }
 }