+++ src/main/java/guideme/internal/search/EnglishAnalyzer.java
@@ -9,6 +9,9 @@ import org.apache.lucene.analysis.en.EnglishPossessiveFilter;
 import org.apache.lucene.analysis.en.KStemFilter;
 import org.apache.lucene.analysis.standard.StandardTokenizer;
 
+/**
+ * Custom version of the default english analyzer to enable the better {@link KStemFilter}.
+ */
 public final class EnglishAnalyzer extends Analyzer {
     @Override
     protected TokenStreamComponents createComponents(String fieldName) {