+++ docs/docs/changelog.md
@@ -14,7 +14,12 @@
   Proguard is enabled to strip unused parts of GuideMEs bundled dependencies (this saves about 5MB). There
   is some potential for `ClassNotFoundErrors` in cases where we missed required Proguard configuration.
   Please let us know if you find any errors!
-
+- API additions
+  - `ConstantColor#TRANSPARENT`
+  - Added `index` method to `TagCompiler` to allow custom tags to control how they are indxed by search.
+    By default, all custom tags simply add their children to the indexer.
+  - Added the ability to set borders for `LytBox`.
+  - Generalized `GuideUiHost` into `DocumentUiHost`.
 ## 2.1.2
 
 - Skip fully invisible blocks (without block entities) when calculating the bounding box of a game scene. Fixes inexplicably larger bounds when blocks like `minecraft:light` where included in the exported structure.